+++ src/main/java/modelo/Egreso.java
@@ -7,6 +7,6 @@ public interface Egreso{
 	public DocumentoComercial getDocumentoComercial();	
 	public Date getFechaEgreso();
 	public MedioDePago getMedioDePagoEgreso();
-	public Double getTotalEgreso();
+	public float getTotalEgreso();
 	public String getDetalleEgreso();
 }
